feat: rework album module to interact with database

+++ src/album/album.service.ts
@@ -1,16 +1,12 @@
-import {
-  Inject,
-  Injectable,
-  NotFoundException,
-  UnprocessableEntityException,
-  forwardRef,
-} from '@nestjs/common';
-import { v4 as uuidV4 } from 'uuid';
+import { Inject, Injectable, NotFoundException, forwardRef } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
 
 import { CreateAlbumDto, UpdateAlbumDto } from './dto';
 import { TrackService } from 'src/track/track.service';
 import { FavsService } from 'src/favs/favs.service';
-import { ArtistService } from 'src/artist/artist.service';
+
+import { Album } from './entities/album.entity';
 
 @Injectable()
 export class AlbumService {
@@ -21,78 +17,45 @@ export class AlbumService {
     @Inject(forwardRef(() => FavsService))
     private readonly favsService: FavsService,
 
-    @Inject(forwardRef(() => ArtistService))
-    private readonly artistService: ArtistService,
+    @InjectRepository(Album)
+    private albumRepository: Repository<Album>,
   ) {}
 
-  private albums: { [id: string]: Album } = {};
-
-  private findAlbum(id: string): Album {
-    const album = this.albums[id];
+  private async findAlbum(id: string): Promise<Album> {
+    const album = await this.albumRepository.findOneBy({ id });
     if (!album) throw new NotFoundException('Album not found');
 
     return album;
   }
 
-  private async checkArtistExistence(artistId: string | null) {
-    if (artistId) {
-      try {
-        await this.artistService.findOne(artistId);
-      } catch {
-        throw new UnprocessableEntityException(`Artist with id ${artistId} not found`);
-      }
-    }
+  create(createAlbumDto: CreateAlbumDto) {
+    const createdAlbum = this.albumRepository.create(createAlbumDto);
+
+    return this.albumRepository.save(createdAlbum);
   }
 
-  async create(createAlbumDto: CreateAlbumDto) {
-    const id = uuidV4();
-    const { artistId, name, year } = createAlbumDto;
-
-    await this.checkArtistExistence(artistId);
-
-    const newAlbum: Album = { artistId, id, name, year };
-
-    this.albums[id] = newAlbum;
-
-    return newAlbum;
+  async findAll() {
+    return (await this.albumRepository.find()).map((album) => album.toResponse());
   }
 
-  findAll() {
-    return Object.values(this.albums);
-  }
-
-  findOne(id: string) {
-    return this.findAlbum(id);
+  async findOne(id: string) {
+    return (await this.findAlbum(id)).toResponse();
   }
 
   async update(id: string, updateAlbumDto: UpdateAlbumDto) {
-    this.findAlbum(id);
+    const album = await this.findAlbum(id);
 
-    const { artistId, name, year } = updateAlbumDto;
+    Object.assign(album, updateAlbumDto);
 
-    await this.checkArtistExistence(artistId);
-
-    const updatedAlbum: Album = { artistId, id, name, year };
-
-    this.albums[id] = updatedAlbum;
-
-    return updatedAlbum;
+    return this.albumRepository.save(album);
   }
 
-  remove(id: string) {
-    this.findAlbum(id);
-    this.albums = Object.fromEntries(Object.entries(this.albums).filter(([key]) => key !== id));
+  async remove(id: string) {
+    const deleteResult = await this.albumRepository.delete(id);
+
+    if (!deleteResult.affected) throw new NotFoundException('Album not found');
+
     this.trackService.removeAlbumFromTracks(id);
     this.favsService.removeAlbum(id);
-  }
-
-  removeArtistFromAlbums(artistId: string) {
-    this.albums = Object.fromEntries(
-      Object.entries(this.albums).map(([albumId, album]) => {
-        if (album.artistId === artistId) return [albumId, { ...album, artistId: null }];
-
-        return [albumId, album];
-      }),
-    );
   }
 }
